+++ src/settings/Locale.js
@@ -30,7 +30,7 @@ const Locale = (props) => {
     // e.g. given the current locale is `ar` and the keys `ar` and `zh-CN` show:
     //    العربية / العربية
     //    الصينية (الصين) / 中文（中国）
-    locale.label = `${intl.formatDisplayName(locale.value)} / ${locale.intl.formatDisplayName(locale.value)}`;
+    locale.label = `${intl.formatDisplayName(locale.value, { type: 'language' })} / ${locale.intl.formatDisplayName(locale.value, { type: 'language' })}`;
   });
 
   const setLocale = (locale) => {