+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Update to react-intl v5. UID-25.
 * Fix calls to `FormatDisplayName()` to work with `react-intl`. Fixes UID-30.
 * Import supported locales from `@folio/stripes/core`.
+* Update queries related to OKAPI-863, OKAPI-835. Fixes UID-34.
 
 ## [3.1.0](https://github.com/folio-org/ui-developer/tree/v3.1.0) (IN PROGRESS)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v3.0.0...v3.1.0)