+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for ui-developer
 
+## 4.1.0 IN PROGRESS
+
 ## [4.0.0](https://github.com/folio-org/ui-developer/tree/v4.0.0) (2020-10-09)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v3.0.0...v4.0.0)